move parse code to utilities
