+++ client-course-schedulizer/src/utilities/interfaces/dataInterfaces.ts
